Restructured folders and fixed errors
